fix some failing tests
