+++ Zpp/Test/Configuration/TestConfigurationFileNames.cs
@@ -5,15 +5,12 @@ namespace Zpp.Test.Configuration
         private const string _basePath = "../../../Test/Configuration/";
         
         // desk
-        public const string DESK_COP_6_LOTSIZE_2 = _basePath + "tisch_cop_6_lotsize_2.json";
-        
         public const string DESK_COP_5_SEQUENTIALLY_LOTSIZE_2 = _basePath + "tisch_cop_5_sequentially_lotsize_2.json";
         public const string DESK_COP_5_CONCURRENT_LOTSIZE_2 = _basePath + "tisch_cop_5_concurrent_lotsize_2.json";
         
         public const string DESK_COP_1_LOTSIZE_1 = _basePath + "tisch_cop_1_lotsize_1.json";
         public const string DESK_COP_1_LOT_ORDER_QUANTITY = _basePath + "tisch_cop_1_lot_order_quantity.json";
         // truck
-        public const string TRUCK_COP_6_LOTSIZE_2 = _basePath + "truck_cop_6_lotsize_2.json";
         public const string TRUCK_COP_5_LOTSIZE_2 = _basePath + "truck_cop_5_lotsize_2.json";
         public const string TRUCK_COP_1_LOTSIZE_1 = _basePath + "truck_cop_1_lotsize_1.json";
     }
