--- Zpp/Test/Configuration/tisch_cop_6_lotsize_2.json
@@ -1,7 +1,0 @@
-{
-  "CustomerOrderPartQuantity": 6,
-  "LotSize": 2,
-  "Name": "desk_cop_6_lotsize_2",
-  "DbSetInitializer": "Zpp.Test.Configuration.MasterDBInitializerDesk",
-  "TestScenario": "Zpp.Test.Configuration.Scenarios.DeskScenarioSequentially"
-}
