+++ Zpp/Test/Configuration/truck_cop_1_lotsize_1.json
@@ -2,6 +2,6 @@
   "CustomerOrderPartQuantity": 1,
   "LotSize": 1,
   "Name": "truck_cop_1_lotsize_1",
-  "DbSetInitializer": "Master40.DB.Data.Initializer.MasterDBInitializerSmall, Master40.DB",
+  "DbSetInitializer": "Master40.DB.Data.Initializer.MasterDBInitializerTruck, Master40.DB",
   "TestScenario": "Zpp.Test.Configuration.Scenarios.TruckScenario"
 }
