--- Zpp/Test/Configuration/truck_cop_6_lotsize_2.json
@@ -1,7 +1,0 @@
-{
-  "CustomerOrderPartQuantity": 6,
-  "LotSize": 2,
-  "Name": "truck_cop_6_lotsize_2",
-  "DbSetInitializer": "Master40.DB.Data.Initializer.MasterDBInitializerSmall, Master40.DB",
-  "TestScenario": "Zpp.Test.Configuration.Scenarios.TruckScenario"
-}
