+++ Zpp/Test/Integration_Tests/TestScheduling.cs
@@ -196,7 +196,7 @@ namespace Zpp.Test.Integration_Tests
             }
         }
         
-        [Theory(Skip = "TODO: this is not correct implemented, so skip it for now.")]
+        [Theory]
         [InlineData(TestConfigurationFileNames.DESK_COP_1_LOT_ORDER_QUANTITY)]
         [InlineData(TestConfigurationFileNames.DESK_COP_5_CONCURRENT_LOTSIZE_2)]
         [InlineData(TestConfigurationFileNames.DESK_COP_5_SEQUENTIALLY_LOTSIZE_2)]
